requirements.txt
streamlit
google-generativeai
requests
pandas
Pillow